+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Inner zoom image to the product image.
+- Add _Login_ Component.
 
 ### Changed
 - Slightly changed the layout of `ProductName` and `SkuSelector` to accomplish proposed design.